Adds further structure to Lagoon Drutiny Problem parser

+++ services/webhooks2tasks/src/handlers/problems/processDrutinyResults.js
@@ -14,29 +14,50 @@ const {
   getEnvironmentByName,
 } = require('@lagoon/commons/src/api');
 
+const ERROR_STATES = ["error", "failure"];
+const SEVERITY_LEVELS = [
+  "NONE",
+  "UNKNOWN",
+  "NEGLIGIBLE",
+  "LOW",
+  "MEDIUM",
+  "HIGH",
+  "CRITICAL"
+];
+const DEFAULT_SEVERITY_LEVEL = "UNKNOWN";
+
 async function processDrutinyResultset(
   webhook: WebhookRequestData,
   channelWrapperWebhooks
 ) {
-  const { webhooktype, event, uuid, body } = webhook;
 
+  const { webhooktype, event, uuid, body } = webhook;
   const { lagoonInfo, results, profile: drutinyProfile } = body;
 
   try {
     const lagoonProjectName =
-      lagoonInfo.LAGOON_SAFE_PROJECT || lagoonInfo.LAGOON_PROJECT;
+      lagoonInfo.LAGOON_DRUTINY_PROJECT_NAME || lagoonInfo.LAGOON_PROJECT || lagoonInfo.LAGOON_SAFE_PROJECT;
 
-    let { id: lagoonProjectId } = await getProjectByName(lagoonProjectName);
+    if(!lagoonProjectName) {
+      throw new Error("No project name passed in Drutiny results");
+    }
 
-    let { environmentByName: environmentDetails } = await getEnvironmentByName(
-      lagoonInfo.LAGOON_ENVIRONMENT,
+    const lagoonEnvironmentName = lagoonInfo.LAGOON_DRUTINY_ENVIRONMENT_NAME || lagoonInfo.LAGOON_ENVIRONMENT || lagoonInfo.LAGOON_GIT_BRANCH;
+
+    if(!lagoonEnvironmentName) {
+      throw new Error("No environment name passed in Drutiny results");
+    }
+
+    const { id: lagoonProjectId } = await getProjectByName(lagoonProjectName);
+
+    const { environmentByName: environmentDetails } = await getEnvironmentByName(
+      lagoonEnvironmentName,
       lagoonProjectId
     );
 
-    let lagoonEnvironmentName = environmentDetails.name;
-    let lagoonEnvironmentId = environmentDetails.id;
-    let lagoonServiceName = DRUTINY_SERVICE_NAME;
-    let drutinyVulnerabilitySource = `${DRUTINY_VULNERABILITY_SOURCE_BASE}`;
+    const lagoonEnvironmentId = environmentDetails.id;
+    const lagoonServiceName = DRUTINY_SERVICE_NAME;
+    const drutinyVulnerabilitySource = `${DRUTINY_VULNERABILITY_SOURCE_BASE}-${drutinyProfile}`;
 
     await deleteProblemsFromSource(
       environmentDetails.id,
@@ -46,17 +67,16 @@ async function processDrutinyResultset(
 
     if (results) {
       results
-        .filter((e) => e.type == 'error')
+        .filter((e) => ERROR_STATES.includes(e.type))
         .forEach((element) => {
           addProblem({
             environment: lagoonEnvironmentId,
             identifier: element.name,
-            severity: element.severity.toUpperCase(),
+            severity: convertSeverityLevels(element.severity),
             source: drutinyVulnerabilitySource,
             description: element.description,
             data: JSON.stringify(element),
             service: DRUTINY_SERVICE_NAME,
-            package: drutinyProfile,
           })
             .then(() => {
               sendToLagoonLogs(
@@ -70,7 +90,7 @@ async function processDrutinyResultset(
                   lagoonEnvironmentId,
                   lagoonEnvironmentName,
                   lagoonServiceName,
-                  severity: element.severity.toUpperCase(),
+                  severity: convertSeverityLevels(element.severity),
                   vulnerability: element,
                 },
                 `New problem found for ${lagoonProjectName}:${lagoonEnvironmentName}:${lagoonServiceName}. Severity: ${element.severity}. Description: ${element.description}`
@@ -98,6 +118,15 @@ async function processDrutinyResultset(
       `Could not process incoming Drutiny scan results, reason: ${error}`
     );
   }
+}
+
+const convertSeverityLevels = (level) => {
+  level = level.toUpperCase();
+  if(SEVERITY_LEVELS.includes(level)) {
+    return level;
+  }
+
+  return DEFAULT_SEVERITY_LEVEL;
 }
 
 module.exports = processDrutinyResultset;
